refactor: 정의한 클래스(Movies) 주입 및 불필요한 코드 제거
