+++ src/components/MovieItem.ts
@@ -1,5 +1,5 @@
 import { POSTER_BASE_URL } from '../api';
-import { MovieInfo } from '../api/api-type';
+import { MovieInfo } from '../domain/Movies';
 import { NO_IMAGE, STAR_FILLED } from '../resource';
 
 const MovieItem = ({ title, poster_path, vote_average }: MovieInfo) => {
